requirements.txt
infi.systray
PIL
requests